refactor google-meet-integration to return single transcript relates #11
